[UPDATE] AccessTokenException 처리 완료
