+++ src/main/java/com/server/EZY/exception/ExceptionAdviceImpl.java
@@ -1,0 +1,65 @@
+package com.server.EZY.exception;
+
+import com.server.EZY.response.ResponseService;
+import com.server.EZY.response.result.CommonResult;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.MessageSource;
+import org.springframework.context.i18n.LocaleContextHolder;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@Slf4j
+@RestControllerAdvice
+@RequiredArgsConstructor
+public class ExceptionAdviceImpl implements ExceptionAdvice{
+
+    private final ResponseService responseService;
+    private final MessageSource messageSource;
+
+    // code 정보에 해당하는 메시지를 조회한다.
+    private String getMessage(String code){
+        return getMessage(code, null);
+    }
+
+    // code 정보, 추가 argument로 현재 locale에 맞는 메시지를 조회합니다.
+    private String getMessage(String code, Object[] args){
+        return messageSource.getMessage(code, args, LocaleContextHolder.getLocale());
+    }
+
+    /**
+     * Exception 처리시 i18n애 정의 되어있는 exception code, msg를 가져온다.
+     * @param code exception 식별 이름
+     * @return CommonResult - 실패 객체
+     */
+    private CommonResult getExceptionResponseObj(String code){
+        return responseService.getFailResult(Integer.valueOf(getMessage(code + ".code")), getMessage(code + ".msg"));
+    }
+
+    // 알수없는 에러
+    @Override
+    public CommonResult defaultException(Exception ex){
+        log.info("=== UnknownException 발생 === \n{}", ex.getMessage());
+        CommonResult exceptionResponseObj = getExceptionResponseObj("user-not-found");
+        exceptionResponseObj.setMassage(
+                exceptionResponseObj.getMassage() + " 원인: \n" + ex.getMessage()
+        );
+        return getExceptionResponseObj("unknown");
+    }
+
+    //*** UserException ***//
+
+    // 유저를 찾을 수 없습니다.
+    @Override
+    public CommonResult userNotFoundException(Exception ex){
+        log.debug("=== User Not Found Exception 발생 ===");
+        return getExceptionResponseObj("user-not-found");
+    }
+
+    @Override
+    public CommonResult accessTokenExpiredException(Exception ex) {
+        log.debug("=== Access Token Expired Exception 발생 ===");
+        return getExceptionResponseObj("access-token-expired");
+    }
+
+
+}
