+++ src/main/resources/i18n/exception_ko.yml
@@ -1,7 +1,11 @@
 unknown:
   code: "-9999"
-  msg: "알 수 없는 에러가 발생했습니다. 서버에 문의해주세요"
+  msg: "알 수 없는 오류가 발생했습니다."
 
 user-not-found:
   code: "-900"
   msg: "해당 유저를 찾을 수 없습니다."
+
+access-token-expired:
+  code : "-800"
+  msg: "엑세스 토큰이 만료되었습니다."
